pyproject.toml
[project]
name = "manual-sih-rag"
version = "2.1.0"
description = "Servico de auditoria de faturamento hospitalar SIH/SUS com RAG e DATASUS"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.24.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "rich>=13.0.0",
    "beautifulsoup4>=4.12.0",
    "google-genai>=1.0.0",
    "rank-bm25>=0.2.2",
    "mcp[cli]>=1.0.0",
    "boto3>=1.28.0",
    "pyarrow>=14.0.0",
    "duckdb>=1.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
py-modules = ["consulta_manual", "extrair_manual", "validar_critica", "analisar_critica", "busca_hibrida", "validar_resposta", "mcp_server", "s3_client", "sigtap_client", "cnes_client"]

[project.scripts]
manual-sih = "consulta_manual:main"
manual-sih-mcp = "mcp_server:main"
manual-sih-server = "mcp_server:main_server"
